+++ requirements-dev.txt
@@ -1,5 +1,5 @@
 depthai-sdk==1.13.1
-viam-sdk==0.11.0
+viam-sdk==0.25.2
 google-api-python-client==2.0.0
 pytest
 black
