+++ requirements.txt
@@ -1,4 +1,4 @@
 depthai-sdk==1.13.1
-viam-sdk==0.11.0
+viam-sdk==0.25.2
 google-api-python-client==2.0.0
 pyinstaller==6.3.0
